Moved code from node.h to node.c
