+++ node.c
@@ -1,4 +1,33 @@
-int add(int n1, int n2)
+#include <stdlib.h>
+#include <string.h>
+
+typedef struct Node{
+  struct Node* next;
+  char* compound;
+} Node;
+
+void insert(Node** head, char* newCompound)
 {
-  return n1 + n2;
+  Node* newNode = malloc(sizeof(Node));
+  newNode->compound = malloc(sizeof(char) * strlen(newCompound));
+  strcpy(newNode->compound, newCompound);
+  newNode->next = *head;
+  *head = newNode;
+}
+
+int listLength(Node** head)
+{
+  Node* current = *head;
+  if (!current)
+  {
+    return 0;
+  }
+  int numElements = 0;
+  while (current)
+  {
+    current = current->next;
+    numElements++;
+  }
+
+  return numElements;
 }
